+++ docs/site/content/en/docs/overview/changelog.md
@@ -22,6 +22,8 @@ You even can **subscribe to get new updates** there!
 ## 2023.10
 October 2023
 - Add `message.revoked` event in **WEBJS** engine
+- Add `me` field in webhook payload - now you can know which number is associated with the session
+- Add `enviroment` field in webhook payload - now you can know which environment is used for the session
 
 ## 2023.9
 September 2023